added arduino source files
right now this directory can't be used in the Arduino IDE (something to
due with Arduino being picky about folder location), but it has all the
necessary code.

Significant differences from the original RF24 code include setting the
SPI clock to 8MHz in RF24.cpp, and setting the radio data rate to 2Mbps.